+++ packages/admin-ui/src/app/core/components/notification/notification.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, ElementRef, HostListener, ViewChild } from '@angular/core';
+import { Component, ElementRef, HostListener, ViewChild } from '@angular/core';
 
 import { NotificationType } from '../../providers/notification/notification.service';