+++ packages/admin-ui/src/app/shared/components/extension-host/extension-host-config.ts
@@ -1,0 +1,10 @@
+export interface ExtensionHostOptions {
+    extensionUrl: string;
+}
+
+export class ExtensionHostConfig {
+    public extensionUrl: string;
+    constructor(options: ExtensionHostOptions) {
+        this.extensionUrl = options.extensionUrl;
+    }
+}