+++ packages/admin-ui/src/compiler/watch.ts
@@ -6,6 +6,8 @@ import * as path from 'path';
 
 import {
     copyExtensionModules,
+    copyStaticAsset,
+    copyUiDevkit,
     createExtensionsModules,
     deleteExistingExtensionModules,
     getModuleOutputDir,
@@ -28,14 +30,16 @@ export function watchAdminUiApp(extensions: Array<Required<AdminUiExtension>>, p
     restoreExtensionsModules();
     deleteExistingExtensionModules();
     copyExtensionModules(extensions);
+    copyUiDevkit();
     createExtensionsModules(extensions);
 
     const config = isInVendureMonorepo() ? 'plugin-dev' : 'plugin';
-    const buildProcess = spawn('yarn', ['ng', 'serve', `-c=${config}`, `--port=${port}`], {
+    const buildProcess = spawn('yarn', ['ng', 'serve', `-c=${config}`, `--port=${port}`, `--poll=1000`], {
         cwd,
         shell: true,
         stdio: 'inherit',
     });
+    const devkitPath = require.resolve('@vendure/ui-devkit');
 
     let watcher: FSWatcher | undefined;
     for (const extension of extensions) {
@@ -50,13 +54,29 @@ export function watchAdminUiApp(extensions: Array<Required<AdminUiExtension>>, p
     }
 
     if (watcher) {
+        // watch the ui-devkit package files too
+        watcher.add(devkitPath);
+    }
+
+    if (watcher) {
         watcher.on('change', filePath => {
             const extension = extensions.find(e => filePath.includes(e.extensionPath));
             if (extension) {
+                if (extension.staticAssets) {
+                    for (const assetPath of extension.staticAssets) {
+                        if (filePath.includes(assetPath)) {
+                            copyStaticAsset(assetPath);
+                            return;
+                        }
+                    }
+                }
                 const outputDir = getModuleOutputDir(extension);
                 const filePart = path.relative(extension.extensionPath, filePath);
                 const dest = path.join(outputDir, filePart);
                 fs.copyFile(filePath, dest);
+            }
+            if (filePath.includes(devkitPath)) {
+                copyUiDevkit();
             }
         });
     }