+++ scripts/docs/generate-typescript-docs.ts
@@ -24,6 +24,7 @@ const sections: DocsSectionConfig[] = [
             'packages/email-plugin/src/',
             'packages/elasticsearch-plugin/src/',
             'packages/testing/src/',
+            'packages/ui-devkit/src/',
         ],
         exclude: [
             /generated-shop-types/,